+++ lib/core.dart
@@ -8,4 +8,4 @@ library;
 export 'package:llm_dart_ai/llm_dart_ai.dart';
 
 export 'core/cancellation.dart'
-    show TransportCancellation, TransportCancelledException;
+    show CancellationHelper, TransportCancellation, TransportCancelledException;